requirements.txt
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pydub==0.25.1
Pygments==2.19.2
pytest==8.4.2
requests==2.32.5
sniffio==1.3.1
urllib3==2.5.0
websocket-client==1.8.0
